Fixed logic and float problems

+++ countit_neon_arduino_sketch/light_control.h
@@ -23,11 +23,11 @@
 // 10- Initializing state turn on all ---> channels 2,3,4,5,6
 
 class TurnOn {
-  float score;
+  int score;
   int* arr = new int[5];
 
   public:
-      TurnOn(float value)
+      TurnOn(int value)
       {
         score = value;
       }
